--- src/__init__.py
@@ -1,6 +1,0 @@
-import logging
-
-from .types import *
-from .types import __all__ as all_types
-
-__all__ = [].extend(all_types)
