+++ src/spotify-to-tidal/sync.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-from . import *
+# from . import *
 
 import logging
 import typing
