+++ src/spotify-to-tidal/type/playlist.py
@@ -1,6 +1,5 @@
-from typing import Union
-
 from tidalapi.playlist import Playlist, UserPlaylist
+from typing import Union
 
 
 TidalPlaylist = Union[Playlist, UserPlaylist]
